fix: Royal Jelly gain read as 0 on huge runs — jelly is u64, gain computed in f64
gainFromReset clamped the f32 gain to maxInt(u32) converted to f32, which
rounds up to 2^32: out of u32 range, so in ReleaseFast the @intFromFloat
was UB and read as 0. Any run past ~4.6e22 honey (Refinery 10) showed
"Royal Jelly gained +0" and the Ascend button stayed disabled.

Royal Jelly (lifetime and spent) is now u64 — lifetime jelly was one
prestige from the u32 ceiling anyway — the gain is computed in f64 with a
representable clamp (1e18), and non-finite honey yields 0. Save lines
parse as u64 (older values load unchanged). Regression test covers the
reported run.

+++ src/save.zig
@@ -60,11 +60,11 @@ pub const Data = struct {
     beehive_upgrade_cost: f32 = 20,
     grid_width: usize = 17,
     grid_height: usize = 17,
-    royal_jelly: u32 = 0,
+    royal_jelly: u64 = 0,
     this_run_honey: f32 = 0,
     prestige_unlocked: bool = false,
     /// Royal Jelly spent in the prestige shop (absent in older saves).
-    jelly_spent: u32 = 0,
+    jelly_spent: u64 = 0,
     /// Prestige shop level per item, indexed by prestige.ShopItem.
     shop_levels: [MAX_SHOP_ITEMS]u16 = [_]u16{0} ** MAX_SHOP_ITEMS,
     // Legacy labs line. aura_multiplier is derived from tree levels on load;
@@ -265,7 +265,7 @@ pub fn read(allocator: std.mem.Allocator, io: std.Io, save_path: []const u8) !Da
             data.grid_height = try parse(usize, tokens.next());
         } else if (std.mem.eql(u8, key, "prestige")) {
             saw_prestige = true;
-            data.royal_jelly = try parse(u32, tokens.next());
+            data.royal_jelly = try parse(u64, tokens.next());
             data.this_run_honey = try parse(f32, tokens.next());
             data.prestige_unlocked = (try parse(u8, tokens.next())) != 0;
         } else if (std.mem.eql(u8, key, "labs")) {
@@ -275,7 +275,7 @@ pub fn read(allocator: std.mem.Allocator, io: std.Io, save_path: []const u8) !Da
             data.burst_cooldown = try parse(f32, tokens.next());
             data.bloom_cooldown = try parse(f32, tokens.next());
         } else if (std.mem.eql(u8, key, "jelly_spent")) {
-            data.jelly_spent = try parse(u32, tokens.next());
+            data.jelly_spent = try parse(u64, tokens.next());
         } else if (std.mem.eql(u8, key, "shop")) {
             const id = try parse(usize, tokens.next());
             const lvl = try parse(u16, tokens.next());
@@ -412,8 +412,8 @@ test "save data survives an atomic round trip" {
     try std.testing.expectEqual(@as(f32, 0.25), restored.fx_volume);
     try std.testing.expectEqual(@as(f32, 9_000_000), restored.honey);
     try std.testing.expectEqual(@as(f32, 128), restored.beehive_factor);
-    try std.testing.expectEqual(@as(u32, 42), restored.royal_jelly);
-    try std.testing.expectEqual(@as(u32, 17), restored.jelly_spent);
+    try std.testing.expectEqual(@as(u64, 42), restored.royal_jelly);
+    try std.testing.expectEqual(@as(u64, 17), restored.jelly_spent);
     try std.testing.expectEqual(@as(u16, 3), restored.shop_levels[2]);
     try std.testing.expectEqual(@as(u16, 0), restored.shop_levels[0]);
     try std.testing.expectEqual(@as(f64, 2.5e12), restored.stats.lifetimeHoney);
@@ -465,7 +465,7 @@ test "flower lines without the is_super field still parse (old saves)" {
     try std.testing.expectEqual(@as(f32, 0.4), restored.music_volume);
     try std.testing.expectEqual(@as(f32, 0.4), restored.fx_volume);
     // No shop lines: nothing spent, nothing owned.
-    try std.testing.expectEqual(@as(u32, 0), restored.jelly_spent);
+    try std.testing.expectEqual(@as(u64, 0), restored.jelly_spent);
     try std.testing.expectEqual(@as(u16, 0), restored.shop_levels[0]);
     try std.testing.expectEqual(@as(usize, 1), restored.flowers.items.len);
     try std.testing.expect(!restored.flowers.items[0].is_super);
